% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALISTA_ordering_main.R
\name{CALISTA_ordering_main}
\alias{CALISTA_ordering_main}
\title{A calista Function}
\usage{
CALISTA_ordering_main(DATA, Results, cell_assignments)
}
\description{
A calista Function
}
\examples{


CALISTA_ordering_main()
}
\keyword{calista}
