% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CALISTA_transition_genes_main.R
\name{CALISTA_transition_genes_main}
\alias{CALISTA_transition_genes_main}
\title{A calista Function}
\usage{
CALISTA_transition_genes_main(DATA, INPUTS, Results, cell_assignments)
}
\description{
A calista Function
}
\examples{


 CALISTA_transition_genes_main()
}
\keyword{calista}
