% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckNumberOfEdges.R
\name{CheckNumberOfEdges}
\alias{CheckNumberOfEdges}
\title{A calista Function}
\usage{
CheckNumberOfEdges(i, MaxNumberOfEdges)
}
\description{
A calista Function
}
\examples{
CheckNumberOfEdge()
}
\keyword{calista}
