% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_distance_GUI.R
\name{cluster_distance_GUI}
\alias{cluster_distance_GUI}
\title{cluster_distance_GUI infer lineage progression among cell clusters}
\usage{
cluster_distance_GUI(DATA, INPUTS, Results)
}
\description{
CALISTA uses the cluster distances - a measure of cluster-cluster
 dissimilarity - to infer the lineage progression or cluster-cluster
 relationship. The lineage progression graph is built based on adding
 edges between clusters in increasing magnitude of cluster distance.
 CALISTA also provides a simple user-interface to add and remove edges
 based on the cluster distances.
}
\details{
Usage:
 Run CALISTA lineage inference using CALISTA single-clustering result
 Results=CALISTA_transition_main(DATA,Results);

Run CALISTA lineage Inference with user-defined cell clusters
A calista Function
}
\examples{
cluster_distance_GUI()
}
\keyword{calista}
