% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_progression2.R
\name{find_progression2}
\alias{find_progression2}
\title{A calista Function}
\usage{
find_progression2(Results, DATA, method)
}
\description{
A calista Function
}
\examples{
find_progression2()
}
\keyword{calista}
