% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prob_transition_genes.R
\name{get_prob_transition_genes}
\alias{get_prob_transition_genes}
\title{A calista Function}
\usage{
get_prob_transition_genes(selected_clusters, log_p_mat_ma, temp_mRNA_all,
  n_genes, nvar_temp)
}
\description{
A calista Function
}
\examples{

get_prob_transistion_genes()
}
\keyword{calista}
