% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_cabsel_f.R
\name{greedy_cabsel_f}
\alias{greedy_cabsel_f}
\title{A calista Function}
\usage{
greedy_cabsel_f(as_all, log_p_mat_ma, k_new, mRNA_all, n_genes, max_iter,
  nvars, optimize, p, sum_prob_tot, population, loops, expected_clusters,
  algorithm, display, opt_idx_a)
}
\description{
A calista Function
}
\examples{

greedy_cabsel_f()
}
\keyword{calista}
