% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_cabsel_ff.R
\name{greedy_cabsel_ff}
\alias{greedy_cabsel_ff}
\title{A calista Function}
\usage{
greedy_cabsel_ff(as_all, log_p_mat_ma, K_new, mRNA_all, n_genes, max_iter,
  nvars, optimize, opt_idx_a, p, sum_prob_tot, population, loops,
  expected_clusters, algorithm, display)
}
\description{
A calista Function
}
\examples{

greedy_cabsel_ff()
}
\keyword{calista}
