% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization}
\alias{normalization}
\title{A calista Function}
\usage{
normalization(data_type, format_data, perczeros_genes, perczeros_cells,
  cut_variable_genes, cells_2_cut)
}
\description{
A calista Function
}
\examples{
normalization()
}
\keyword{calista}
