% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_4_calista_ordering.R
\name{plot_4_calista_ordering}
\alias{plot_4_calista_ordering}
\title{A calista Function}
\usage{
plot_4_calista_ordering(groups, colorMARK, group_names, Results)
}
\description{
A calista Function
}
\examples{
plot_4_calista_ordering()
}
\keyword{calista}
