% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_cluster_GUI.R
\name{reorder_clusters_GUI}
\alias{reorder_clusters_GUI}
\title{A calista Function}
\usage{
reorder_clusters_GUI(DATA, INPUTS, Results)
}
\description{
A calista Function
}
\examples{
reorder_clusters_GUI()
}
\keyword{calista}
