% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopQuietly_calista.R
\name{stopQuietly_calista}
\alias{stopQuietly_calista}
\title{A calista Function}
\usage{
stopQuietly_calista(...)
}
\description{
A calista Function
}
\examples{
stopQuietly_calista()
}
\keyword{calista}
