% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_RmEdge.R
\name{temp_RmEdge}
\alias{temp_RmEdge}
\title{A calista Function}
\usage{
temp_RmEdge(h, node, colorMARK2, x_center, y_center, z_center,
  ClusterGroup2, expected_clusters, normed_pseudotime, score3, legendInfo,
  az, el)
}
\description{
A calista Function
}
\examples{
temp_RmEdge()
}
\keyword{calista}
