% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_transition_GUI.R
\name{update_transition_GUI}
\alias{update_transition_GUI}
\title{A calista Function}
\usage{
update_transition_GUI(DATA, INPUTS, Results, nodes, edges_chosen)
}
\description{
A calista Function
}
\examples{
update_transition_GUI()
}
\keyword{calista}
